backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pymupdf==1.24.0
openai>=1.40.0
supabase==2.5.0
python-dotenv==1.0.0
pydantic>=2.7.0
httpx>=0.27.0
tiktoken>=0.7.0
deprecation==2.1.0
